Book.javaの内容をJoinedBook.javaに変更。JoinedBook.java使用箇所をBook.javaに変更
