+++ src/main/java/com/ookawara/book/application/entity/Book.java
@@ -1,20 +1,20 @@
 package com.ookawara.book.application.entity;
 
-import java.time.LocalDate;
-
 public class Book {
     private int book_id;
     private String name;
-    private LocalDate release_date;
-    private boolean is_purchased;
+    private String release_date;
+    private Boolean is_purchased;
     private int category_id;
+    private String category;
 
-    public Book(int book_id, String name, LocalDate release_date, boolean is_purchased, int category_id) {
+    public Book(int book_id, String name, String release_date, Boolean is_purchased, int category_id, String category) {
         this.book_id = book_id;
         this.name = name;
         this.release_date = release_date;
         this.is_purchased = is_purchased;
         this.category_id = category_id;
+        this.category = category;
     }
 
     public int getBookId() {
@@ -25,7 +25,7 @@ public class Book {
         return name;
     }
 
-    public LocalDate getReleaseDate() {
+    public String getReleaseDate() {
         return release_date;
     }
 
@@ -35,5 +35,9 @@ public class Book {
 
     public int getCategoryId() {
         return category_id;
+    }
+
+    public String getCategory() {
+        return category;
     }
 }
