+++ src/test/java/com/ookawara/book/application/BookApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class JoinedBookApplicationTests {
+class BookApplicationTests {
 
 	@Test
 	void contextLoads() {
